GME Plugin: Silence a type truncation warning
Adding a cast here silences a warning from passing a long to a function
accepting an int. It doesn't really matter here anyway, as the long in
question is hard coded to initialize to a fixed sample rate. Even when
sample rate configuration is eventually added, this will still be hard
capped to well within the range of 32-bit integers.

+++ Plugins/GME/GameDecoder.m
@@ -58,7 +58,7 @@ gme_err_t readCallback(void *data, void *out, int count) {
 	if(type == gme_spc_type || type == gme_sfm_type)
 		sampleRate = 32000;
 
-	emu = gme_new_emu(type, sampleRate);
+	emu = gme_new_emu(type, (int)sampleRate);
 	if(!emu) {
 		ALog(@"GME: No new emu!");
 		return NO;
